refactor: Move code
